requirements.txt
# LiveKit Agents framework (latest stable versions)
livekit-agents>=1.2.0
livekit-plugins-openai>=1.2.0
livekit-plugins-deepgram>=1.2.0
livekit-plugins-elevenlabs>=0.6.0

# Core dependencies
python-dotenv==1.0.0
aiohttp>=3.9.1
pytz>=2023.3
